Create sass-watch workflow
default calls init. init calls in sequence sass, inject and scss watch. This compiles the scss to css, injects it into index.html and starts a watch on changes .scss files.
It handles file events and reacts differently to the type of change:
when a .scss is deleted, the file its corresponding .css file is removes from the app/css folder. It's link in index.html is also removed. Setting the option empty to true in gulp-injector allows the inject:css section to be empty if the css folder is empty.
if a file is added, it compiles it to css and injects it.
if a file is changed, it compiles it to css and injects it.