Remove clean:dist task. Replace with array of clean tasks. Add array of watch tasks.
clean:dist was running clean:js and clean:css in sequence, however, the main sequence from where clean:dist was being called was not waiting for clean:dist to be done before calling the next task, sass. This async process was causing errors when running the gulp command consecutively. sass was being run before clean:css and all that it had to work with was an empty folder.