Use ng-finder to better sort and organize angular files
Since ng-finder gathers the existing .js files, there is no need to recreate the angular source on changes -- the files are the same. The source only needs to be rebuit when a .js file is deleted or added. The path to .css needs to be added to the angular source as ng-finder only returns angular files.