Watch over deletion and creation of .html files
Prior to this change, the html-watch only watched over changes to .html files. From a SPA architecture perspective, this made sense since index.html only changes - it is never created or destroyed. However, what if a template is renamed after the .js file where the template is reference to is updated with the new template name? The browser would not reload and console errors will prevail. With the new event listeners, the watcher reloads the browser when an .html file is renamed. The respective .js file will be able to pull the template behind doors and trigger the refresh.